test(refactor): rename test helpers folder
