+++ ciphering-cli-tool/test/cli-helpers.spec.js
@@ -1,4 +1,4 @@
-const { mockArgv } = require('./lib/mock-argv')
+const { mockArgv } = require('./__helpers/mock-argv')
 
 const {
   isInDebugMode,
