+++ ciphering-cli-tool/test/transform/atbash-transformer.spec.js
@@ -1,6 +1,6 @@
 const { Transform } = require('stream')
 
-const { testTransformer } = require('../lib/test-transformer')
+const { testTransformer } = require('../__helpers/test-transformer')
 
 const { AtbashTransformer } = require('../../lib/transform/atbash-transformer')
