+++ ciphering-cli-tool/test/transform/rot-transformer.spec.js
@@ -3,7 +3,7 @@ const { Transform } = require('stream')
 const matchers = require('jest-extended')
 expect.extend(matchers);
 
-const { testTransformer } = require('../lib/test-transformer')
+const { testTransformer } = require('../__helpers/test-transformer')
 
 const { RotTransformer } = require('../../lib/transform/rot-transformer')
 const { DIRECTION_ } = RotTransformer
